refactor(cmd): move expand helpers from create.go to expand.go (#5781)
