+++ cmd/create_test.go
@@ -24,84 +24,6 @@ import (
 	. "gopkg.in/check.v1"
 )
 
-func (s *MySuite) TestSetCLIVariables(c *C) {
-	ds := config.DeploymentSettings{
-		Vars: config.Dict{}.
-			With("deployment_name", cty.StringVal("bush"))}
-
-	vars := []string{
-		"project_id=cli_test_project_id",
-		"deployment_name=cli_deployment_name",
-		"region=cli_region",
-		"zone=cli_zone",
-		"kv=key=val",
-		"keyBool=true",
-		"keyInt=15",
-		"keyFloat=15.43",
-		"keyMap={bar: baz, qux: 1}",
-		"keyArray=[1, 2, 3]",
-		"keyArrayOfMaps=[foo, {bar: baz, qux: 1}]",
-		"keyMapOfArrays={foo: [1, 2, 3], bar: [a, b, c]}",
-	}
-	c.Assert(setCLIVariables(&ds, vars), IsNil)
-	c.Check(
-		ds.Vars.Items(), DeepEquals, map[string]cty.Value{
-			"project_id":      cty.StringVal("cli_test_project_id"),
-			"deployment_name": cty.StringVal("cli_deployment_name"),
-			"region":          cty.StringVal("cli_region"),
-			"zone":            cty.StringVal("cli_zone"),
-			"kv":              cty.StringVal("key=val"),
-			"keyBool":         cty.True,
-			"keyInt":          cty.NumberIntVal(15),
-			"keyFloat":        cty.NumberFloatVal(15.43),
-			"keyMap": cty.ObjectVal(map[string]cty.Value{
-				"bar": cty.StringVal("baz"),
-				"qux": cty.NumberIntVal(1)}),
-			"keyArray": cty.TupleVal([]cty.Value{
-				cty.NumberIntVal(1), cty.NumberIntVal(2), cty.NumberIntVal(3)}),
-			"keyArrayOfMaps": cty.TupleVal([]cty.Value{
-				cty.StringVal("foo"),
-				cty.ObjectVal(map[string]cty.Value{
-					"bar": cty.StringVal("baz"),
-					"qux": cty.NumberIntVal(1)})}),
-			"keyMapOfArrays": cty.ObjectVal(map[string]cty.Value{
-				"foo": cty.TupleVal([]cty.Value{
-					cty.NumberIntVal(1), cty.NumberIntVal(2), cty.NumberIntVal(3)}),
-				"bar": cty.TupleVal([]cty.Value{
-					cty.StringVal("a"), cty.StringVal("b"), cty.StringVal("c")}),
-			}),
-		})
-
-	// Failure: Variable without '='
-	ds = config.DeploymentSettings{}
-	inv := []string{"project_idcli_test_project_id"}
-	c.Check(setCLIVariables(&ds, inv), ErrorMatches, "invalid format: .*")
-
-	// Failure: Unmarshalable value
-	ds = config.DeploymentSettings{}
-	inv = []string{"pyrite={gold"}
-	c.Check(setCLIVariables(&ds, inv), ErrorMatches, ".*unable to convert.*pyrite.*gold.*")
-}
-
-func (s *MySuite) TestSetBackendConfig(c *C) {
-	// Success
-	vars := []string{
-		"taste=sweet",
-		"type=green",
-		"odor=strong",
-	}
-
-	ds := config.DeploymentSettings{}
-	c.Assert(setBackendConfig(&ds, vars), IsNil)
-
-	be := ds.TerraformBackendDefaults
-	c.Check(be.Type, Equals, "green")
-	c.Check(be.Configuration.Items(), DeepEquals, map[string]cty.Value{
-		"taste": cty.StringVal("sweet"),
-		"odor":  cty.StringVal("strong"),
-	})
-}
-
 func (s *MySuite) TestMergeDeploymentSettings(c *C) {
 	ds1 := config.DeploymentSettings{
 		Vars: config.Dict{}.
@@ -164,49 +86,6 @@ func (s *MySuite) TestMergeDeploymentSettings(c *C) {
 			"bucket": cty.StringVal("ds_bucket"),
 		}),
 	})
-}
-
-func (s *MySuite) TestSetBackendConfig_Invalid(c *C) {
-	// Failure: Variable without '='
-	vars := []string{
-		"typegreen",
-	}
-	ds := config.DeploymentSettings{}
-	c.Assert(setBackendConfig(&ds, vars), ErrorMatches, "invalid format: .*")
-}
-
-func (s *MySuite) TestSetBackendConfig_NoOp(c *C) {
-	ds := config.DeploymentSettings{
-		TerraformBackendDefaults: config.TerraformBackend{
-			Type: "green"}}
-
-	c.Assert(setBackendConfig(&ds, []string{}), IsNil)
-	c.Check(ds.TerraformBackendDefaults, DeepEquals, config.TerraformBackend{
-		Type: "green"})
-}
-
-func (s *MySuite) TestValidationLevels(c *C) {
-	bp := config.Blueprint{}
-
-	c.Check(setValidationLevel(&bp, "ERROR"), IsNil)
-	c.Check(bp.ValidationLevel, Equals, config.ValidationError)
-
-	c.Check(setValidationLevel(&bp, "WARNING"), IsNil)
-	c.Check(bp.ValidationLevel, Equals, config.ValidationWarning)
-
-	c.Check(setValidationLevel(&bp, "IGNORE"), IsNil)
-	c.Check(bp.ValidationLevel, Equals, config.ValidationIgnore)
-
-	c.Check(setValidationLevel(&bp, "INVALID"), NotNil)
-}
-
-func (s *MySuite) TestValidateMaybeDie(c *C) {
-	bp := config.Blueprint{
-		Validators:      []config.Validator{{Validator: "invalid"}},
-		ValidationLevel: config.ValidationWarning,
-	}
-	ctx, _ := config.NewYamlCtx([]byte{})
-	validateMaybeDie(bp, ctx) // smoke test
 }
 
 func (s *MySuite) TestIsOverwriteAllowed_Absent(c *C) {
